Save states: Removed MS-specific extension usage & disabled extensions + Removed some errors/warnings caused by this change

+++ Core/Breakpoint.cpp
@@ -12,7 +12,7 @@ Breakpoint::~Breakpoint()
 
 bool Breakpoint::Matches(uint32_t memoryAddr, uint32_t absoluteAddr)
 {
-	return _addr == -1 || (memoryAddr == _addr && !_isAbsoluteAddr) || (absoluteAddr == _addr && _isAbsoluteAddr);
+	return _addr == -1 || ((int32_t)memoryAddr == _addr && !_isAbsoluteAddr) || ((int32_t)absoluteAddr == _addr && _isAbsoluteAddr);
 }
 
 BreakpointType Breakpoint::GetType()
